Externalize Commons TOSCA Types Store them only once in a unique location
