+++ data/tosca/yorc-kubernetes-types.yml
@@ -3,7 +3,7 @@ tosca_definitions_version: yorc_tosca_simple_yaml_1_0
 metadata:
   template_name: yorc-kubernetes-types
   template_author: yorc
-  template_version: 1.0.0
+  template_version: 2.0.0
 
 imports:
   - normative: <yorc-docker-types.yml>
