+++ data/tosca/yorc-slurm-types.yml
@@ -3,7 +3,7 @@ tosca_definitions_version: yorc_tosca_simple_yaml_1_0
 metadata:
   template_name: yorc-slurm-types
   template_author: yorc
-  template_version: 1.0.0
+  template_version: 1.2.0
 
 imports:
   - yorc: <yorc-types.yml>
