+++ deployments/internal/store_requirements.go
@@ -1,0 +1,33 @@
+// Copyright 2018 Bull S.A.S. Atos Technologies - Bull, Rue Jean Jaures, B.P.68, 78340, Les Clayes-sous-Bois, France.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package internal
+
+import (
+	"net/url"
+
+	"github.com/ystia/yorc/v3/helper/consulutil"
+	"github.com/ystia/yorc/v3/tosca"
+)
+
+func StoreRequirementAssignment(consulStore consulutil.ConsulStore, requirement tosca.RequirementAssignment, requirementPrefix, requirementName string) {
+	consulStore.StoreConsulKeyAsString(requirementPrefix+"/name", requirementName)
+	consulStore.StoreConsulKeyAsString(requirementPrefix+"/node", requirement.Node)
+	consulStore.StoreConsulKeyAsString(requirementPrefix+"/relationship", requirement.Relationship)
+	consulStore.StoreConsulKeyAsString(requirementPrefix+"/capability", requirement.Capability)
+	consulStore.StoreConsulKeyAsString(requirementPrefix+"/type_requirement", requirement.TypeRequirement)
+	for propName, propValue := range requirement.RelationshipProps {
+		StoreValueAssignment(consulStore, requirementPrefix+"/properties/"+url.QueryEscape(propName), propValue)
+	}
+}
