+++ deployments/operations.go
@@ -88,7 +88,10 @@ func GetOperationPathAndPrimaryImplementation(kv *api.KV, deploymentID, nodeTemp
 		typeOrNodeTemplate = nodeTemplateImpl
 	}
 
-	operationPath := getOperationPath(deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	operationPath, err := getOperationPath(kv, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	if err != nil {
+		return "", "", errors.Wrapf(err, "Failed to retrieve primary implementation for operation %q on template/type %q", operationName, typeOrNodeTemplate)
+	}
 	kvp, _, err := kv.Get(path.Join(operationPath, "implementation/primary"), nil)
 	if err != nil {
 		return "", "", errors.Wrapf(err, "Failed to retrieve primary implementation for operation %q on template/type %q", operationName, typeOrNodeTemplate)
@@ -123,7 +126,10 @@ func GetOperationPathAndPrimaryImplementation(kv *api.KV, deploymentID, nodeTemp
 
 func getOperationPathAndPrimaryImplementationForNodeType(kv *api.KV, deploymentID, nodeType, operationName string) (string, string, error) {
 	// First check if operation exists in current nodeType
-	operationPath := getOperationPath(deploymentID, "", nodeType, operationName)
+	operationPath, err := getOperationPath(kv, deploymentID, "", nodeType, operationName)
+	if err != nil {
+		return "", "", err
+	}
 	importPath, err := GetTypeImportPath(kv, deploymentID, nodeType)
 	if err != nil {
 		return "", "", err
@@ -155,14 +161,14 @@ func getOperationPathAndPrimaryImplementationForNodeType(kv *api.KV, deploymentI
 }
 
 // This function return the path for a given operation
-func getOperationPath(deploymentID, nodeTemplate, nodeType, operationName string) string {
-	opPath, _ := getOperationAndInterfacePath(deploymentID, nodeTemplate, nodeType, operationName)
-	return opPath
+func getOperationPath(kv *api.KV, deploymentID, nodeTemplate, nodeType, operationName string) (string, error) {
+	opPath, _, err := getOperationAndInterfacePath(kv, deploymentID, nodeTemplate, nodeType, operationName)
+	return opPath, err
 }
 
 // This function return the path for a given operation and the path of its interface
 // It handles the ways that implementation is a node template or a node type
-func getOperationAndInterfacePath(deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName string) (string, string) {
+func getOperationAndInterfacePath(kv *api.KV, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName string) (string, string, error) {
 	var operationPath, interfacePath string
 	opShortName := stringutil.GetLastElement(operationName, ".")
 	interfaceName := stringutil.GetAllExceptLastElement(operationName, ".")
@@ -170,11 +176,15 @@ func getOperationAndInterfacePath(deploymentID, nodeTemplateImpl, nodeTypeImpl, 
 		operationPath = path.Join(consulutil.DeploymentKVPrefix, deploymentID, "topology/nodes", nodeTemplateImpl, "interfaces", interfaceName, opShortName)
 		interfacePath = path.Join(consulutil.DeploymentKVPrefix, deploymentID, "topology/nodes", nodeTemplateImpl, "interfaces", interfaceName)
 	} else {
-		operationPath = path.Join(consulutil.DeploymentKVPrefix, deploymentID, "topology/types", nodeTypeImpl, "interfaces", interfaceName, opShortName)
-		interfacePath = path.Join(consulutil.DeploymentKVPrefix, deploymentID, "topology/types", nodeTypeImpl, "interfaces", interfaceName)
+		typePath, err := locateTypePath(kv, deploymentID, nodeTypeImpl)
+		if err != nil {
+			return "", "", err
+		}
+		operationPath = path.Join(typePath, "interfaces", interfaceName, opShortName)
+		interfacePath = path.Join(typePath, "interfaces", interfaceName)
 	}
 
-	return operationPath, interfacePath
+	return operationPath, interfacePath, nil
 }
 
 // GetRelationshipTypeImplementingAnOperation  returns the first (bottom-up) type in the type hierarchy of a given relationship that implements a given operation
@@ -187,7 +197,10 @@ func GetRelationshipTypeImplementingAnOperation(kv *api.KV, deploymentID, nodeNa
 	}
 	relType := relTypeInit
 	for relType != "" {
-		operationPath := getOperationPath(deploymentID, "", relType, operationName)
+		operationPath, err := getOperationPath(kv, deploymentID, "", relType, operationName)
+		if err != nil {
+			return "", err
+		}
 		keys, _, err := kv.Keys(operationPath+"/", "/", nil)
 		if err != nil {
 			return "", errors.Wrap(err, consulutil.ConsulGenericErrMsg)
@@ -215,7 +228,10 @@ func GetNodeTypeImplementingAnOperation(kv *api.KV, deploymentID, nodeName, oper
 
 // IsNodeTemplateImplementingOperation returns true if the node implements the defined operation
 func IsNodeTemplateImplementingOperation(kv *api.KV, deploymentID, nodeName, operationName string) (bool, error) {
-	operationPath := getOperationPath(deploymentID, nodeName, "", operationName)
+	operationPath, err := getOperationPath(kv, deploymentID, nodeName, "", operationName)
+	if err != nil {
+		return false, errors.Wrapf(err, "Can't define if operation with name:%q exists for node %q", operationName, nodeName)
+	}
 	keys, _, err := kv.Keys(operationPath+"/", "/", nil)
 	if err != nil {
 		return false, errors.Wrapf(err, "Can't define if operation with name:%q exists for node %q", operationName, nodeName)
@@ -234,7 +250,10 @@ func GetTypeImplementingAnOperation(kv *api.KV, deploymentID, typeName, operatio
 	implType := typeName
 	for implType != "" {
 		log.Debugf("[GetTypeImplementingAnOperation] implType=%q", implType)
-		operationPath := getOperationPath(deploymentID, "", implType, operationName)
+		operationPath, err := getOperationPath(kv, deploymentID, "", implType, operationName)
+		if err != nil {
+			return "", err
+		}
 		log.Debugf("[GetTypeImplementingAnOperation] operationPath=%q", operationPath)
 
 		keys, _, err := kv.Keys(operationPath+"/", "/", nil)
@@ -251,7 +270,10 @@ func GetTypeImplementingAnOperation(kv *api.KV, deploymentID, typeName, operatio
 
 // GetOperationImplementationType allows you when the implementation of an operation is an artifact to retrieve the type of this artifact
 func GetOperationImplementationType(kv *api.KV, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName string) (string, error) {
-	operationPath := getOperationPath(deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	operationPath, err := getOperationPath(kv, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	if err != nil {
+		return "", errors.Wrap(err, "Fail to get the type of operation implementation")
+	}
 	kvp, _, err := kv.Get(path.Join(operationPath, "implementation/type"), nil)
 	if err != nil {
 		return "", errors.Wrap(err, "Fail to get the type of operation implementation")
@@ -294,7 +316,10 @@ func GetOperationImplementationType(kv *api.KV, deploymentID, nodeTemplateImpl, 
 //
 // The returned file is the raw value. To have a file with a path relative to the root of the deployment use GetOperationImplementationFileWithRelativePath()
 func GetOperationImplementationFile(kv *api.KV, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName string) (string, error) {
-	operationPath := getOperationPath(deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	operationPath, err := getOperationPath(kv, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	if err != nil {
+		return "", errors.Wrap(err, "Fail to get the file of operation implementation")
+	}
 	kvp, _, err := kv.Get(path.Join(operationPath, "implementation/file"), nil)
 	if err != nil {
 		return "", errors.Wrap(err, "Fail to get the file of operation implementation")
@@ -326,7 +351,10 @@ func GetOperationImplementationFileWithRelativePath(kv *api.KV, deploymentID, no
 
 // GetOperationImplementationRepository allows you when the implementation of an operation is an artifact to retrieve the repository of this artifact
 func GetOperationImplementationRepository(kv *api.KV, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName string) (string, error) {
-	operationPath := getOperationPath(deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	operationPath, err := getOperationPath(kv, deploymentID, nodeTemplateImpl, nodeTypeImpl, operationName)
+	if err != nil {
+		return "", errors.Wrap(err, "Fail to get the file of operation implementation")
+	}
 	kvp, _, err := kv.Get(path.Join(operationPath, "implementation/repository"), nil)
 	if err != nil {
 		return "", errors.Wrap(err, "Fail to get the file of operation implementation")
@@ -443,7 +471,10 @@ func GetImplementationArtifactForOperation(kv *api.KV, deploymentID, nodeName, o
 
 // GetOperationInputs returns the list of inputs names for a given operation
 func GetOperationInputs(kv *api.KV, deploymentID, nodeTemplateImpl, typeNameImpl, operationName string) ([]string, error) {
-	operationPath, interfacePath := getOperationAndInterfacePath(deploymentID, nodeTemplateImpl, typeNameImpl, operationName)
+	operationPath, interfacePath, err := getOperationAndInterfacePath(kv, deploymentID, nodeTemplateImpl, typeNameImpl, operationName)
+	if err != nil {
+		return nil, err
+	}
 	// First Get operation inputs
 	inputKeys, _, err := kv.Keys(operationPath+"/inputs/", "/", nil)
 	if err != nil {
@@ -511,7 +542,10 @@ func GetOperationInput(kv *api.KV, deploymentID, nodeName string, operation prov
 		return nil, errors.Errorf("Input %q for operation %v is a property definition we can't resolve it without a task input", inputName, operation)
 	}
 
-	operationPath, interfacePath := getOperationAndInterfacePath(deploymentID, operation.ImplementedInNodeTemplate, operation.ImplementedInType, operation.Name)
+	operationPath, interfacePath, err := getOperationAndInterfacePath(kv, deploymentID, operation.ImplementedInNodeTemplate, operation.ImplementedInType, operation.Name)
+	if err != nil {
+		return nil, err
+	}
 	inputPath := path.Join(operationPath, "inputs", inputName, "data")
 	res, isFunction, err := getValueAssignmentWithoutResolve(kv, deploymentID, inputPath, "")
 	if err != nil {
@@ -620,7 +654,10 @@ func GetOperationInputPropertyDefinitionDefault(kv *api.KV, deploymentID, nodeNa
 	} else if !isPropDef {
 		return nil, errors.Errorf("Input %q for operation %v is not a property definition we can't resolve its default value", inputName, operation)
 	}
-	operationPath, interfacePath := getOperationAndInterfacePath(deploymentID, operation.ImplementedInNodeTemplate, operation.ImplementedInType, operation.Name)
+	operationPath, interfacePath, err := getOperationAndInterfacePath(kv, deploymentID, operation.ImplementedInNodeTemplate, operation.ImplementedInType, operation.Name)
+	if err != nil {
+		return nil, err
+	}
 	inputPath := path.Join(operationPath, "inputs", inputName, "default")
 	// TODO base datatype should be retrieved
 	res, isFunction, err := getValueAssignmentWithoutResolve(kv, deploymentID, inputPath, "")
@@ -669,7 +706,10 @@ func GetOperationInputPropertyDefinitionDefault(kv *api.KV, deploymentID, nodeNa
 
 // IsOperationInputAPropertyDefinition checks if a given operation input is a property definition
 func IsOperationInputAPropertyDefinition(kv *api.KV, deploymentID, nodeTemplateImpl, typeNameImpl, operationName, inputName string) (bool, error) {
-	operationPath, interfacePath := getOperationAndInterfacePath(deploymentID, nodeTemplateImpl, typeNameImpl, operationName)
+	operationPath, interfacePath, err := getOperationAndInterfacePath(kv, deploymentID, nodeTemplateImpl, typeNameImpl, operationName)
+	if err != nil {
+		return false, err
+	}
 	kvp, _, err := kv.Get(path.Join(operationPath, "inputs", inputName, "is_property_definition"), nil)
 	if err != nil {
 		return false, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
@@ -710,7 +750,10 @@ func GetOperationHostFromTypeOperation(kv *api.KV, deploymentID, typeName, inter
 // The returned value may be an empty string. This function doesn't explore the type heirarchy to
 // find the operation or declared value for operation_host.
 func GetOperationHostFromTypeOperationByName(kv *api.KV, deploymentID, typeName, operationName string) (string, error) {
-	opPath, _ := getOperationAndInterfacePath(deploymentID, "", typeName, operationName)
+	opPath, _, err := getOperationAndInterfacePath(kv, deploymentID, "", typeName, operationName)
+	if err != nil {
+		return "", err
+	}
 	hop := path.Join(opPath, "implementation/operation_host")
 	kvp, _, err := kv.Get(hop, nil)
 	if err != nil || kvp == nil {
@@ -724,7 +767,10 @@ func GetOperationHostFromTypeOperationByName(kv *api.KV, deploymentID, typeName,
 // An implemented operation means that it has a non empty primary implementation or file for an implementation artifact
 func IsOperationImplemented(kv *api.KV, deploymentID, nodeName, operationName string) (bool, error) {
 	// First check on node template
-	nodeTemplateOpPath := getOperationPath(deploymentID, nodeName, "", strings.ToLower(operationName))
+	nodeTemplateOpPath, err := getOperationPath(kv, deploymentID, nodeName, "", strings.ToLower(operationName))
+	if err != nil {
+		return false, err
+	}
 	result, err := isOperationPathImplemented(kv, nodeTemplateOpPath)
 	if err != nil {
 		return false, err
@@ -739,7 +785,10 @@ func IsOperationImplemented(kv *api.KV, deploymentID, nodeName, operationName st
 		return false, err
 	}
 	for typeName != "" {
-		typeNameOpPath := getOperationPath(deploymentID, "", typeName, strings.ToLower(operationName))
+		typeNameOpPath, err := getOperationPath(kv, deploymentID, "", typeName, strings.ToLower(operationName))
+		if err != nil {
+			return false, err
+		}
 		result, err := isOperationPathImplemented(kv, typeNameOpPath)
 		if err != nil {
 			return false, err
