+++ server/builtin_types_testing.go
@@ -1,0 +1,8 @@
+// +build testing
+
+package server
+
+func registerBuiltinTOSCATypes() error {
+	// In testing those types are registered by the testutil package
+	return nil
+}
