+++ testutil/helper.go
@@ -57,6 +57,9 @@ func NewTestConsulInstance(t *testing.T) (*testutil.TestServer, *api.Client) {
 
 	kv := client.KV()
 	consulutil.InitConsulPublisher(cfg.Consul.PubMaxRoutines, kv)
+
+	storeBuiltinDefinitions()
+
 	return srv1, client
 }
